core: porting: restructure and update comm files based on ui change
