+++ src/core/mgr_task_executor.c
@@ -1,0 +1,150 @@
+/**
+ * @file mgr_task_executor.c
+ *
+ */
+
+/*********************
+ *      INCLUDES
+ *********************/
+// #define LOG_LEVEL LOG_LEVEL_TRACE
+#if defined(LOG_LEVEL)
+#warning "LOG_LEVEL defined locally will override the global setting in this file"
+#endif
+#include <log.h>
+
+#include <stdlib.h>
+#include <stdint.h>
+
+#include <comm/dbus_comm.h>
+#include <comm/cmd_payload.h>
+#include <sched/workqueue.h>
+#include <sched/task.h>
+#include <devices/imu.h>
+
+/*********************
+ *      DEFINES
+ *********************/
+
+/**********************
+ *      TYPEDEFS
+ **********************/
+
+/**********************
+ *  GLOBAL VARIABLES
+ **********************/
+
+/**********************
+ *  STATIC PROTOTYPES
+ **********************/
+
+/**********************
+ *  STATIC VARIABLES
+ **********************/
+
+/**********************
+ *      MACROS
+ **********************/
+
+/**********************
+ *   STATIC FUNCTIONS
+ **********************/
+
+/**********************
+ *   GLOBAL FUNCTIONS
+ **********************/
+int32_t process_opcode_endless(uint32_t opcode, void *data)
+{
+    int32_t ret = 0;
+
+    switch (opcode) {
+    case OP_START_DBUS:
+        ret = dbus_fn_thread_handler();
+        break;
+    case OP_START_IMU:
+        ret = imu_kalman_init("/sys/bus/iio/devices/iio:device2/", 100, 0.001f, 0.003f, 0.03f);
+        imu_kalman_set_debug(1);
+        if (!ret)
+            ret = imu_fn_thread_handler();
+        break;
+    default:
+        LOG_ERROR("Opcode [%d] is invalid", opcode);
+        break;
+    }
+
+    return ret;
+}
+
+int32_t process_opcode(uint32_t opcode, void *data)
+{
+    int32_t ret = 0;
+
+    switch (opcode) {
+    case OP_BACKLIGHT_INIT:
+        backlight_setup();
+        break;
+    case OP_BACKLIGHT_DEINIT:
+        break;
+    case OP_SET_BRIGHTNESS:
+        set_brightness( (*((remote_cmd_t *)data)).entries[1].value.i32);
+        break;
+    case OP_GET_BRIGHTNESS:
+        break;
+    case OP_LEFT_VIBRATOR:
+        ret = rumble_trigger(2, 80, 150);
+        break;
+    case OP_RIGHT_VIBRATOR:
+        ret = rumble_trigger(3, 80, 150);
+        break;
+    case OP_STOP_IMU:
+        imu_fn_thread_stop();
+        break;
+    case OP_READ_IMU:
+        struct imu_angles a = imu_get_angles();
+        LOG_DEBUG("roll=%.2f pitch=%.2f yaw=%.2f\n", a.roll, a.pitch, a.yaw);
+        break;
+    case OP_AUDIO_INIT:
+        snd_sys_init();
+        break;
+    case OP_AUDIO_RELEASE:
+        snd_sys_release();
+        break;
+    case OP_SOUND_PLAY:
+        // TODO: support sound file path
+        audio_play_sound("/usr/share/sounds/sound-icons/percussion-10.wav");
+        break;
+
+    default:
+        LOG_ERROR("Opcode [%d] is invalid", opcode);
+        break;
+    }
+
+    return ret;
+}
+
+int32_t create_local_simple_task(uint8_t flow, uint8_t duration, uint32_t opcode)
+{
+    work_t *work = create_work(LOCAL, flow, duration, opcode, NULL);
+    if (!work) {
+        LOG_ERROR("Failed to create work from cmd");
+        return -1;
+    }
+
+    push_work(work);
+
+    return 0;
+}
+
+int32_t create_remote_task(uint8_t flow, void *data)
+{
+    work_t *work;
+
+    work = create_work(REMOTE, flow, SHORT, OP_DBUS_SENT_CMD_DATA, data);
+    if (!work) {
+        LOG_ERROR("Failed to create work from cmd");
+        return -1;
+    }
+
+    push_work(work);
+
+    return 0;
+}
