+++ src/core/task/workqueue.c
@@ -1,16 +1,46 @@
+/**
+ * @file workqueue.c
+ *
+ */
+
+/*********************
+ *      INCLUDES
+ *********************/
+// #define LOG_LEVEL LOG_LEVEL_TRACE
+#if defined(LOG_LEVEL)
+#warning "LOG_LEVEL defined locally will override the global setting in this file"
+#endif
+#include <log.h>
+
 #include <stdlib.h>
 #include <stdint.h>
 #include <pthread.h>
 #include <signal.h>
 
-#include <dbus_comm.h>
+#include <comm/dbus_comm.h>
+#include <sched/workqueue.h>
+#include <sched/task.h>
 
-#include <workqueue.h>
-#include <task.h>
-#include <log.h>
+/*********************
+ *      DEFINES
+ *********************/
 
+/**********************
+ *      TYPEDEFS
+ **********************/
+
+/**********************
+ *  GLOBAL VARIABLES
+ **********************/
 extern volatile sig_atomic_t g_run;
 
+/**********************
+ *  STATIC PROTOTYPES
+ **********************/
+
+/**********************
+ *  STATIC VARIABLES
+ **********************/
 static workqueue_t g_wqueue = {
     .head = NULL,
     .tail = NULL,
@@ -18,39 +48,50 @@ static workqueue_t g_wqueue = {
     .cond = PTHREAD_COND_INITIALIZER
 };
 
+/**********************
+ *      MACROS
+ **********************/
+
+/**********************
+ *   STATIC FUNCTIONS
+ **********************/
+
+/**********************
+ *   GLOBAL FUNCTIONS
+ **********************/
 work_t *create_work(uint8_t type, uint8_t flow, uint8_t duration, \
                     uint32_t opcode, void *data)
 {
-	work_t *w;
+    work_t *w;
 
-	w = calloc(1, sizeof(*w));
-	if (!w)
-		return NULL;
+    w = calloc(1, sizeof(*w));
+    if (!w)
+        return NULL;
 
     w->type = type;
     w->flow = flow;
     w->duration = duration;
     w->opcode = opcode;
-	w->data = data;
-	LOG_TRACE("Created work for opcode: %d", w->opcode);
+    w->data = data;
+    LOG_TRACE("Created work for opcode: %d", w->opcode);
 
-	return w;
+    return w;
 }
 
 void delete_work(work_t *w)
 {
-	if (!w) {
-		LOG_WARN("Unable to delete work: null work pointer");
-		return;
-	}
+    if (!w) {
+        LOG_WARN("Unable to delete work: null work pointer");
+        return;
+    }
 
-	LOG_TRACE("Deleting work for opcode: %d", w->opcode);
-	if (w->data) {
-	    free(w->data);
-		return;
-	}
+    LOG_TRACE("Deleting work for opcode: %d", w->opcode);
+    if (w->data) {
+        free(w->data);
+        return;
+    }
 
-	free(w);
+    free(w);
 }
 
 void push_work(work_t *w) {
